Address PR review: rename custom_disharray -> disharray, remove HIRAX
- Rename drift/telescope/custom_disharray/ to drift/telescope/disharray/,
  replacing the old simple disharray.py module
- Rename CustomDishArray mixin to DishArrayMixin throughout
- Preserve old DishArray class in disharray/core.py for backward compatibility
- Remove HIRAX-specific code (hirax.py, hextile data file, manager.py entries)
  per instrument-agnostic policy; HIRAX implementation belongs in a
  dedicated instrument repo
