+++ drift/telescope/__init__.py
@@ -6,7 +6,6 @@ Telescope class implementation
 
     cylbeam
     cylinder
-    custom_disharray
     disharray
     exotic_cylinder
     focalplane
